+++ src/migrant/tests/test_engine.py
@@ -3,11 +3,16 @@
 # Copyright 2014 by Shoobx, Inc.
 #
 ###############################################################################
+from typing import List, Dict, Generator
+import os
 import unittest
+import time
 
 import mock
 
 from migrant.engine import MigrantEngine
+from migrant.backend import MigrantBackend
+from migrant.repository import Script, Repository
 
 
 class MigrantEngineTest(unittest.TestCase):
@@ -152,6 +157,7 @@ def _make_engine(migrations, scripts, log=None):
     log = log if log is not None else []
     backend = mock.Mock()
     backend.list_migrations.return_value = migrations
+    backend.begin = lambda db: db
 
     backend.generate_test_connections.return_value = ["db1", "db2"]
 
@@ -162,3 +168,113 @@ def _make_engine(migrations, scripts, log=None):
 
     engine = MigrantEngine(backend, repository, {})
     return engine
+
+
+class MultiDbBackend(MigrantBackend[str, str]):
+    def __init__(self, dbs: List[str], logfname: str) -> None:
+        self._applied = {}
+        self.dbs = dbs
+        self.logfname = logfname
+        for db in dbs:
+            self._applied[db] = ["INITIAL"]
+
+    def begin(self, db: str) -> str:
+        return db
+
+    def list_migrations(self, db: str) -> List[str]:
+        return self._applied.get(db, [])
+
+    def push_migration(self, db, migration):
+        migrations = self._applied.setdefault(db, [])
+        migrations.append(migration)
+
+    def pop_migration(self, db, migration):
+        migrations = self._applied.setdefault(db, [])
+        migrations.remove(migration)
+
+    def generate_connections(self) -> Generator[str, None, None]:
+        """Generate connections to process
+        """
+        for db in self.dbs:
+            # with open(self.logfname, "a") as f:
+            #     f.write(f"Opening database: {db}\n")
+            #     print(f"Opening database: {db}")
+            yield db
+
+
+class TimedScript(Script):
+    def __init__(self, name: str, timemap: Dict[str, float], logfname: str) -> None:
+        self.name = name
+        self._timemap = timemap
+        self._logfname = logfname
+
+    def up(self, db):
+        tosleep = self._timemap.get(db, 0)
+        time.sleep(tosleep)
+        with open(self._logfname, "a") as f:
+            f.write(f"{db}: Upgraded to {self.name} ({tosleep}s)\n")
+
+
+class MultiDbRepo(Repository):
+    def __init__(self, timemap: Dict[str, float], logfname: str) -> None:
+        self.timemap = timemap
+        self.logfname = logfname
+
+    def list_script_ids(self) -> List[str]:
+        return ["script1"]
+
+    def load_script(self, scriptid: str) -> Script:
+        return TimedScript(scriptid, self.timemap, self.logfname)
+
+
+def test_concurrent_upgrade_multiprocess(tmp_path) -> None:
+    # GIVEN
+
+    # Prepare the log file. Since we are running scripts in
+    # multiprocessing environment, we cannot share memory data structure, like
+    # list, and let scripts write log to it. Instead, write to a share file.
+    logfname = os.path.join(tmp_path, "migration.log")
+    # backend = MultiDbBackend(["db1"] + ["db2", "db3"]*10, logfname)
+    backend = MultiDbBackend(["db1", "db2", "db3"], logfname)
+    repository = MultiDbRepo({"db1": 0.1, "db2": 0.01, "db3": 0.05}, logfname)
+    engine = MigrantEngine(backend, repository, {}, processes=2)
+
+    # WHEN
+    engine.update()
+
+    # THEN
+    with open(logfname, "r") as f:
+        log = f.read().strip().split("\n")
+
+    # Longest task executed last
+    assert log == [
+        "db2: Upgraded to script1 (0.01s)",
+        "db3: Upgraded to script1 (0.05s)",
+        "db1: Upgraded to script1 (0.1s)",
+    ]
+
+
+def test_concurrent_upgrade_singleprocess(tmp_path) -> None:
+    # GIVEN
+
+    # Prepare the log file. Since we are running scripts in
+    # multiprocessing environment, we cannot share memory data structure, like
+    # list, and let scripts write log to it. Instead, write to a share file.
+    logfname = os.path.join(tmp_path, "migration.log")
+    backend = MultiDbBackend(["db1", "db2", "db3"], logfname)
+    repository = MultiDbRepo({"db1": 0.1, "db2": 0.01, "db3": 0.05}, logfname)
+    engine = MigrantEngine(backend, repository, {}, processes=1)
+
+    # WHEN
+    engine.update()
+
+    # THEN
+    with open(logfname, "r") as f:
+        log = f.read().strip().split("\n")
+
+    # First task executed first despite being the longest
+    assert log == [
+        "db1: Upgraded to script1 (0.1s)",
+        "db2: Upgraded to script1 (0.01s)",
+        "db3: Upgraded to script1 (0.05s)",
+    ]